+++ go.mod
@@ -6,6 +6,13 @@ require (
 	github.com/fatih/color v1.13.0
 	github.com/sirupsen/logrus v1.8.1
 	github.com/vitali-fedulov/images/v2 v2.1.2
+	gorm.io/gorm v1.24.0
+)
+
+require (
+	github.com/jinzhu/inflection v1.0.0 // indirect
+	github.com/jinzhu/now v1.1.5 // indirect
+	github.com/mattn/go-sqlite3 v1.14.15 // indirect
 )
 
 require (
@@ -18,4 +25,5 @@ require (
 	golang.org/x/exp v0.0.0-20221023144134-a1e5550cf13e
 	golang.org/x/sys v0.0.0-20220722155257-8c9f86f7a55f // indirect
 	gopkg.in/yaml.v3 v3.0.1
+	gorm.io/driver/sqlite v1.4.3
 )
